third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/worker-src/shared-fallback.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/testharness-helper.js"></script>
<meta http-equiv="Content-Security-Policy" content="worker-src http://127.0.0.1:8000 blob:; child-src 'none'">
<script>
  var url = new URL("/security/contentSecurityPolicy/resources/ping.js", document.baseURI).toString();
  assert_shared_worker_is_loaded(url, "Same-origin dedicated worker allowed by 'self'.");

  var b = new Blob(["onconnect = e => { e.ports[0].postMessage('ping'); }"], {type: "text/javascript"});
  var url = URL.createObjectURL(b);
  assert_shared_worker_is_loaded(url, "blob: dedicated worker allowed by 'blob:'.");
</script>
